+++ core/Xenode/src/imgui/ImGuiLayer.cpp
@@ -95,7 +95,7 @@ namespace Xen {
 		}
 	}
 
-	void ImGuiLayer::OnUpdate(double timestep)
+	void ImGuiLayer::OnUpdate(float timestep)
 	{
 
 	}
