+++ core/Xenode/src/imgui/ImGuiLayer.h
@@ -27,7 +27,7 @@ namespace Xen {
 
 		virtual void OnAttach() override;
 		virtual void OnDetach() override;
-		virtual void OnUpdate(double timestep) override;
+		virtual void OnUpdate(float timestep) override;
 
 		virtual void OnImGuiUpdate() override;
